ai-service/requirements.txt
fastapi
uvicorn
sentence-transformers
requests
pillow
rembg[cpu]
pydantic